pyproject.toml
[tool.poetry]
name = "nosok-game"
version = "0.1.0"
description = "Simple pygame project"
authors = ["l4blee"]

[tool.poetry.dependencies]
python = "^3.9"
pygame = "^2.1.0"
Pillow = "^9.0.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
